adding a small bit of testing
